+++ CopilotKit/packages/react-textarea/src/types/base/base-autosuggestions-config.tsx
@@ -93,7 +93,7 @@ const defaultShouldAcceptAutosuggestionOnKeyPress = (
   return false;
 };
 
-const defaultShouldAcceptAutosuggestionOnTouch = () => true;
+const defaultShouldAcceptAutosuggestionOnTouch = () => false;
 
 /**
  * Default configuration for the BaseAutosuggestions.