+++ CopilotKit/examples/next-pages-router/CHANGELOG.md
@@ -1,5 +1,28 @@
 # next-pages-router
 
+## 1.0.9
+
+### Patch Changes
+
+- Dev console, bugfixes
+- Updated dependencies
+  - @copilotkit/react-core@1.0.9
+  - @copilotkit/react-textarea@1.0.9
+  - @copilotkit/react-ui@1.0.9
+  - @copilotkit/runtime@1.0.9
+  - @copilotkit/shared@1.0.9
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@1.0.8
+  - @copilotkit/react-textarea@1.0.8
+  - @copilotkit/react-ui@1.0.8
+  - @copilotkit/runtime@1.0.8
+  - @copilotkit/shared@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes