+++ CopilotKit/examples/node-express/CHANGELOG.md
@@ -1,5 +1,22 @@
 # node
 
+## 1.0.9
+
+### Patch Changes
+
+- Dev console, bugfixes
+- Updated dependencies
+  - @copilotkit/runtime@1.0.9
+  - @copilotkit/shared@1.0.9
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/runtime@1.0.8
+  - @copilotkit/shared@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes