+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # ui
 
+## 1.0.9
+
+### Patch Changes
+
+- Dev console, bugfixes
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.0.9
+  - @copilotkit/shared@1.0.9
+
+## 1.0.8
+
+### Patch Changes
+
+- Remove redundant console logs
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.0.8
+  - @copilotkit/shared@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes