+++ CopilotKit/packages/runtime/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @copilotkit/runtime
 
+## 1.0.9
+
+### Patch Changes
+
+- Dev console, bugfixes
+- Updated dependencies
+  - @copilotkit/shared@1.0.9
+
+## 1.0.8
+
+### Patch Changes
+
+- Remove redundant console logs
+- Updated dependencies
+  - @copilotkit/shared@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes